test Comment et topic

+++ Services/CommentService.cs
@@ -27,6 +27,10 @@ namespace forum_api.Services
 
         public Comment DeleteComment(int id)
         {
+            if (id == null)
+            {
+                throw new Exception($"Aucun comment avec l'id {id}, n'a été trouvé.");
+            }
             return _repository.DeleteById(id);
         }
 
@@ -45,13 +49,17 @@ namespace forum_api.Services
             var comment = _repository.FindByTopicId(TpId);
             if (comment == null)
             {
-                throw new Exception($"Aucun topic avec l'id {TpId}, n'a été trouvé.");
+                throw new Exception($"Aucun comment avec l'id {TpId}, n'a été trouvé.");
             }
             return comment;
         }
 
         public Comment UpdateComment(Comment comment)
         {
+            if (comment == null)
+            {
+                throw new Exception($"Mauvais Comment, null.");
+            }
             comment.ModificationDate = DateTime.Now;
             return _repository.UpdateComment(comment);
         }
